perf: optimize leaderboard and enforce analyzers
Consolidate leaderboard scans and resolve analyzer findings across the API, notifications, tests, and UI.

+++ api/DataExplorer/ServerRotationDetector.cs
@@ -138,7 +138,7 @@ public static class ServerRotationDetector
     {
         var cycleCoverage = Math.Min(1.0, (double)candidate.FullCycles / MinimumFullCycles);
         var tailCoverage = Math.Min(1.0, (double)candidate.MatchedRounds / Math.Max(12, sampleSize));
-        var sampleCoverage = Math.Min(1.0, (double)sampleSize / 36.0);
+        var sampleCoverage = Math.Min(1.0, sampleSize / 36.0);
 
         return Math.Round((cycleCoverage * 0.5) + (tailCoverage * 0.3) + (sampleCoverage * 0.2), 2);
     }
